pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "pyfundlib"
dynamic = ["version"]
description = "End-to-End Python Framework for Algorithmic Trading: Backtesting, ML Alpha, Live Execution & Automation"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Himanshu Dixit", email = "hima@pyfund.tech" }]

dependencies = [
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "polars>=1.0.0",
  "scikit-learn>=1.3.0",
  "torch>=2.1.0",
  "xgboost>=2.0.0",
  "mlflow>=2.12.0",
  "optuna>=3.0.0",
  "statsmodels>=0.14.0",
  "scipy>=1.10.0",
  "apscheduler>=3.10.4",
  "yfinance>=0.2.40",
  "pyarrow>=14.0.0",
  "seaborn>=0.13.0",
  "matplotlib>=3.8.0",
  "langchain>=0.1.0",
  "crewai>=0.1.0",
  "openai-whisper",
  "faiss-cpu",
  "structlog",
  "pydantic>=2.0.0",
  "pydantic-settings",
  "elevenlabs",
  "pyttsx3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov",
  "black>=24.0",
  "ruff>=0.6.0",
  "mypy>=1.11",
  "pre-commit",
  "sphinx>=7.0",
  "sphinx-rtd-theme",
  "jupyterlab>=4.0",
  "streamlit>=1.38.0",
]

broker = [
  "alpaca-py==0.28.0",
  "kiteconnect>=4.4.0",
]

full = [
  "pyfundlib[dev]",
  "pyfundlib[broker]",
]

[project.urls]
Homepage = "https://pyfund.tech"
Documentation = "https://pyfundlib.readthedocs.io/en/latest/"
Repository = "https://github.com/Hima-D/pyfundlib"
Issues = "https://github.com/Hima-D/pyfundlib/issues"

[project.scripts]
pyfundlib = "pyfundlib.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pyfundlib"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
strict = true
[tool.ruff]
# Minimal root section (required for ruff format)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "C4", "PIE", "RUF"]
ignore = ["E501"]           # Black owns line length
future-annotations = true

